backend-python/user/requirements.txt
flask
blinker
mysql-connector-python
mysqlclient
flask_sqlalchemy
flask_marshmallow
flask-cors
bcrypt
freeze
pyjwt
elastic-apm[flask]
